refactor: :recycle: Extract functions to open.c
