+++ include/api.h
@@ -16,29 +16,10 @@ void	env_set_exitcode(t_dict *env, int exitcode);
 t_res	env_set(t_dict *env, char *str);
 char	*env_get(t_dict *env, char *key);
 /*
-** < file.c > */
-
-bool	is_dir(char *path);
-bool	is_path(char *path);
-bool	is_executable_exists(char *file, t_dict *env);
-char	*new_executable_from_env(char *file, t_dict *env);
-/*
-** < heredoc.c > */
-
-t_fd	shell_heredoc(t_shell *shell, const char *eof);
-/*
 ** < path.c > */
 
 char	*new_path_resolved(char *path, t_dict *env);
 char	**new_names_from_path(char *name, t_dict *env);
-/*
-** < redirect.c > */
-
-t_res	api_open(t_fd *fd_p, t_AST_NODE *redirect, t_shell *shell);
-void	cmd_try_open_redirect( t_AST_NODE *node, t_AST_COMMAND *cmd,
-			t_shell *shell);
-void	cmd_open_redirects(t_AST_COMMAND *cmd, t_shell *shell);
-void	shell_open_redirects(t_shell *shell);
 /*
 ** < signal.c > */
