+++ src/open/heredoc.c
